search abstract.php also in the directory of the command itself
I invoked the shell script from within a folder different to magento/shell and got a fatal error because the file 'abstract.php' required was not found.

I fixed this by also looking in the directory of the command itself. As no absolut paths are used, this was and still is relative to the current working directory.